requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
yfinance>=0.2
pandas>=2.0
pydantic>=1.10
nest_asyncio>=1.5
vaderSentiment>=3.3
pinecone-client>=2.2
google-genai>=0.1